Cosmetics Store App
This is a cosmetics store app for my perfume business.

The plan is to develop it to be a form of content manager.

Where you can manage the store from within the app. Especially deleting and adding products.

Find a way to save all tokens, and payment details of every transaction in to a database for future.